Move opacity and darkLight into utils
